requirements.txt
async-timeout==4.0.2
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
distlib==0.3.7
filelock==3.12.2
idna==3.4
platformdirs==3.10.0
python-dotenv==1.0.0
redis==4.6.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
spotipy==2.23.0
urllib3==2.0.3
virtualenv==20.24.3
